chore(setup): setup ESLint, Prettier, and Mocha
This commit sets the project up to use ESLint and Prettier. The ESLint
confit is based off of the Airbnb config as well as the Prettier config.
We also setup Mocha to use as a test framework for use with the project.
Along with the use of Mocha we have added our first tests in the form of
a unit test for the 'apiService' of the app.

Partially satisfies: #1
	new file:   .eslintrc
	modified:   package.json
	new file:   test/unit/server/services/apiService.spec.js
	new file:   test/utils/setup-babel-register.js
	modified:   yarn.lock