chore(setup): run ESLint as part of the tests
You can now run ESLint against all of the source/test files in the
project with the 'test:eslint' script. ESLint will also be ran as part
of the tests. Yet it is preferable that you setup async linting so you
don't get caught with a bunch of formatting issues before checking in.

Partially satisfies: #1
	modified:   .gitignore
	modified:   package.json
	modified:   yarn.lock