build(setup): lint commit messages (AngularJS message conventions)
Setup linting for our commit messages. This will ensure that we follow
the AngularJS message conventions. This will be useful in the future if
(when?) we setup automatic releasing of our project with a tool like
'semantic-release'. Conventions for the commit messages can be found
here: https://gist.github.com/stephenparish/9941e89d80e2bc58a153

We also lint our files that are staged with commit. They will be linted
and re-added for submission before our commit message prompt comes up.

Partially satisfies: #1
	modified:   package.json
	modified:   yarn.lock