build(setup): compile server code
This commit compiles the server code to the folder 'lib'. This isn't
entirely useful now but will be in the future for when people want to
standup a server based on this project. You wouldn't want to use
'babel-node' to run the server as of its overhead in resource
consumption.

Closes #1
	modified:   .gitignore
	modified:   package.json
	modified:   yarn.lock