build(setup): complete initial setup of project (#2)
* chore(setup): setup basic server/API service

This commit lays down the foundation for a server that uses an API
service to respond to HTTP requests. In doing so we have taken the
initial steps to start compiling the server code by adding using Babel
to transpile our code.

Partially satisfies: #1
	new file:   .babelrc
	new file:   .gitignore
	new file:   package.json
	new file:   src/server/config.js
	new file:   src/server/honeyBadgerApp.js
	new file:   src/server/server.js
	new file:   src/server/services/apiService.js
	new file:   yarn.lock

* chore(setup): add MIT LICENSE.md

Adding a MIT LICENSED.md file for the project. Also added a link to my
hosted MIT license to the README.

Partially satisfies: #1
	new file:   LICENSE.md
	modified:   README.md

* chore(setup): setup ESLint, Prettier, and Mocha

This commit sets the project up to use ESLint and Prettier. The ESLint
confit is based off of the Airbnb config as well as the Prettier config.
We also setup Mocha to use as a test framework for use with the project.
Along with the use of Mocha we have added our first tests in the form of
a unit test for the 'apiService' of the app.

Partially satisfies: #1
	new file:   .eslintrc
	modified:   package.json
	new file:   test/unit/server/services/apiService.spec.js
	new file:   test/utils/setup-babel-register.js
	modified:   yarn.lock

* chore(setup): run ESLint as part of the tests

You can now run ESLint against all of the source/test files in the
project with the 'test:eslint' script. ESLint will also be ran as part
of the tests. Yet it is preferable that you setup async linting so you
don't get caught with a bunch of formatting issues before checking in.

Partially satisfies: #1
	modified:   .gitignore
	modified:   package.json
	modified:   yarn.lock

* chore(setup): setup ESDoc

ESDoc has been added for the projects use. You can generate API
documentation by running the 'yarn doc' recipe. If you would like start
an API doc server you can run 'yarn start:doc'.
	new file:   .esdoc.json
	modified:   .gitignore
	modified:   package.json
	modified:   src/server/services/apiService.js
	modified:   yarn.lock

* setup(chore): setup CI with Travis

This commit adds the initial Travis CI config file after we turned on
the hooks/integrations with the services in GitHub. We may need to make
subsequent changes after this to complete the Travis CI integration due
to some errors.

Partially satisfies: #1
	new file:   .travis.yml

* build(setup): lint commit messages (AngularJS message conventions)

Setup linting for our commit messages. This will ensure that we follow
the AngularJS message conventions. This will be useful in the future if
(when?) we setup automatic releasing of our project with a tool like
'semantic-release'. Conventions for the commit messages can be found
here: https://gist.github.com/stephenparish/9941e89d80e2bc58a153

We also lint our files that are staged with commit. They will be linted
and re-added for submission before our commit message prompt comes up.

Partially satisfies: #1
	modified:   package.json
	modified:   yarn.lock

* docs: add link/badge to Travis CI builds
	modified:   README.md

* build(setup): compile server code

This commit compiles the server code to the folder 'lib'. This isn't
entirely useful now but will be in the future for when people want to
standup a server based on this project. You wouldn't want to use
'babel-node' to run the server as of its overhead in resource
consumption.

Closes #1
	modified:   .gitignore
	modified:   package.json
	modified:   yarn.lock