feat(gql): implement initial 'honeyBadgers' query (#4)
* feat(gql): implement initial 'honeyBadgers' query

This commit implements our first query named 'honeyBadgers'. The query
allows client consumers to query for the top intruders of a honeypot. As
part of this commit we also setup our schema since it wasn't yet
implemented.

Closes #2

* docs(readme): doc how to setup the app for HoneyDB usage

This commit adds information to the README that explains how to setup
the app for usage with HoneyDB. We use HoneyDB to provide threat
intelligence data about intruders to honey pots.