feat(gql-svc): setup GraphQL API endpoint/service (#6)
This commit creates a GraphQL API endpoint/service. It is
consumed/exposed as Express middleware.

In addition we have added a lot of packages as 'dependencies' that
aren't really dependencies at the moment. I think this is okay as we
will be using them when we begin clientside work soon.

Closes #5