feat(cli): implement initial <App> component, query GraphQL API endpoint
This commit provides the initial implementation of the <App> component
which will server as the topmost component of our app. It makes use of
the Apollo client to query our GraphQL endpoint for the top honey
badgers.

We also added CORS middleware to our Express app. This will allow for
the React hotloader/dev server to serve up the client bundle and allow
for the client to communicate with the dev server while both are
running.

Finally I had to unfortunately increase the timeout of tests that make
use of the GraphQL endpoint. :( My tests were timing out and after
investigating it doesn't seem related to my environment or the fact that
I develop on a VPS. The most likely cause is the latency involved with
querying HoneyDB. I will watch for this though as if we have to increase
our timeout again we may need to think of alternative approaches to
testing.

Partially satisfies #7