feat(cli): setup Material-UI, add unit and integration tests for <App>
This commit makes use of Material-UI. Right now we do little with it but
it is in place for the future. We have also added unit and integration
tests for the <App> component.

We have had to increase the timeout of our tests again to account for
latency with the HoneyDB integration. We need to fix this in the future
as we increase the latency to 15 seconds which is unacceptable. Maybe we
can mock out the data for when we run the tests.

Closes #7