feat(cli): implement initial React UI (#10)
* feat(cli): implement initial <Root> component, bundle with webpack

This commit provides the initial implementation of the <Root> component.
It is trivial and was really created to provide an entry point to our
client bundle for webpack. Spoiler! We also bundle the client with
webpack. :)

Partially satisfies #7

* feat(cli): implement initial <App> component, query GraphQL API endpoint

This commit provides the initial implementation of the <App> component
which will server as the topmost component of our app. It makes use of
the Apollo client to query our GraphQL endpoint for the top honey
badgers.

We also added CORS middleware to our Express app. This will allow for
the React hotloader/dev server to serve up the client bundle and allow
for the client to communicate with the dev server while both are
running.

Finally I had to unfortunately increase the timeout of tests that make
use of the GraphQL endpoint. :( My tests were timing out and after
investigating it doesn't seem related to my environment or the fact that
I develop on a VPS. The most likely cause is the latency involved with
querying HoneyDB. I will watch for this though as if we have to increase
our timeout again we may need to think of alternative approaches to
testing.

Partially satisfies #7

* feat(cli): setup Material-UI, add unit and integration tests for <App>

This commit makes use of Material-UI. Right now we do little with it but
it is in place for the future. We have also added unit and integration
tests for the <App> component.

We have had to increase the timeout of our tests again to account for
latency with the HoneyDB integration. We need to fix this in the future
as we increase the latency to 15 seconds which is unacceptable. Maybe we
can mock out the data for when we run the tests.

Closes #7