feat(gql): refactor 'honeyBadgers' query to 'topHoneyBadgers'
This commit is a refactor to change the 'honeyBadgers' query to
'topHoneyBadgers'. Really I guess this is a breaking change but since we
haven't released yet we won't consider it breaking.

We also include the amount of times an entity was seen in the last 24
hours in a honey pot.

Partially satisfies #9