refactor(gql): refactor 'topHoneyBadgers' query to return top 25
This commit refactors the 'topHoneyBadgers' query to return the top 25
results from HoneyDB instead of the top 100. This was done mostly to
reduce the API "hits"/consumption of the Apility services API. This is
becuase every top result will result in at least 1 request to Apility
for additional info which will eat up your service tiers limits.
	modified:   src/server/graphql/schema.js
	modified:   test/unit/server/graphql/schemaQueries.spec.js