feat(cli): implement table to display data about honey badgers
This commit does just more than implement a table to display data about
honey badgers. It grew in scope as we learned we needed a little more
functionality as well as ways to improve the development experience.

On top of providing a table to display data about honey badgers we
implemented a new GraphQL query that allows one to query for autonomous
systems info associated with a honey badger. This query makes use of the
Apility service/APIs to provide the data allowing us to build up our
threat intelligence data.

Additionally we have provided a way to use mock data from the GraphQL
service/API endpoint instead of querying HoneyDB/Apility directly. This
data is static and won't be useful for realtime analysis. Additionally
it shouldn't be used in production. This functionality was added to make
sure that daily quotas/API usage limits weren't hit from 3rd party
services while doing development. To make use of mock data you can set
the environment variable 'RETURN_MOCK_DATA' to 'true'.

Closes #8
Closes #9
Closes #11