feat(cli): use table to present data and implement new GraphQL query 'autonomousSystems' (#12)
* feat(gql): refactor 'honeyBadgers' query to 'topHoneyBadgers'

This commit is a refactor to change the 'honeyBadgers' query to
'topHoneyBadgers'. Really I guess this is a breaking change but since we
haven't released yet we won't consider it breaking.

We also include the amount of times an entity was seen in the last 24
hours in a honey pot.

Partially satisfies #9

* refactor(gql): refactor 'topHoneyBadgers' query to return top 25

This commit refactors the 'topHoneyBadgers' query to return the top 25
results from HoneyDB instead of the top 100. This was done mostly to
reduce the API "hits"/consumption of the Apility services API. This is
becuase every top result will result in at least 1 request to Apility
for additional info which will eat up your service tiers limits.
	modified:   src/server/graphql/schema.js
	modified:   test/unit/server/graphql/schemaQueries.spec.js

* feat(cli): implement table to display data about honey badgers

This commit does just more than implement a table to display data about
honey badgers. It grew in scope as we learned we needed a little more
functionality as well as ways to improve the development experience.

On top of providing a table to display data about honey badgers we
implemented a new GraphQL query that allows one to query for autonomous
systems info associated with a honey badger. This query makes use of the
Apility service/APIs to provide the data allowing us to build up our
threat intelligence data.

Additionally we have provided a way to use mock data from the GraphQL
service/API endpoint instead of querying HoneyDB/Apility directly. This
data is static and won't be useful for realtime analysis. Additionally
it shouldn't be used in production. This functionality was added to make
sure that daily quotas/API usage limits weren't hit from 3rd party
services while doing development. To make use of mock data you can set
the environment variable 'RETURN_MOCK_DATA' to 'true'.

Closes #8
Closes #9
Closes #11

* fix: don't snapshot dynamic integration tests <App>

The integration test for the React <App> component makes direct usage of
the 3rd party services (HoneyDB and Apility). This fix removes Jest
snapshots from the tests as it will always fail since the data returned
by the 3rd party services is dynamic/realtime and won't match from the
last test run. This ought to be just fine since we do use Jest snapshots
in our unit tests where we can control the data and pass static values
for each test run.

* docs: doc how to setup the app for Apility usage and mock data

This commit updates the README to show how the app can be setup to make
use of Apility. Additionally we show how mock data can be returned by
the GraphQL service/API endpoint so that one doesn't eat into their
limits/quotas for the 3rd party services we use.