feat(gql): add 'geoLocations' query to GQL schema to get locations
This commit introduces a new query to our GraphQL schema named
'geoLocations'. It allows one to supply a list of the IP addresses for a
honey badger and then get the geospatial location (i.e.
latitude/longitude) for the honey badger using that IP address.

In addition ot introducing a new query to our GraphQL schema we have
also updated our <App> component to issue a 'geoLocations' query and our
<HoneyBadgersTable> component to make use of the new data collected.

Closes #13