feat(gql): add details panel for honey badgers, search/filter
This commit implements a details panel that allows for more detailed
information about honey badgers to be displayed. It also adds
functionality to search/fitler honey badgers as well as see a location
of where the honey badger is located in terms of which city they reside
in.

Partially closes #14