feat(gql): add details panel for honey badgers, search/filter
Commit 23bcd5feb32f036562f33dca7af9b6bb9abd20d3 tried to implement a
details panel that allows for more detailed information about honey
badgers to be displayed. Unfortunately I was a dummy and forgot to add
any files to be committed that actually implemented that functionality.
This commit fixes that mistake and actually submits the files that
implements that functionality as well as functionality to search/filter
honey badgers as well as see a location of where the honey badger is
located in terms of which city they reside in.

Partially satisfies #14