feat(cli): implement a Honey Badger profile (#16)
* feat(gql): add details panel for honey badgers, search/filter

This commit implements a details panel that allows for more detailed
information about honey badgers to be displayed. It also adds
functionality to search/fitler honey badgers as well as see a location
of where the honey badger is located in terms of which city they reside
in.

Partially closes #14

* feat(gql): add details panel for honey badgers, search/filter

Commit 23bcd5feb32f036562f33dca7af9b6bb9abd20d3 tried to implement a
details panel that allows for more detailed information about honey
badgers to be displayed. Unfortunately I was a dummy and forgot to add
any files to be committed that actually implemented that functionality.
This commit fixes that mistake and actually submits the files that
implements that functionality as well as functionality to search/filter
honey badgers as well as see a location of where the honey badger is
located in terms of which city they reside in.

Partially satisfies #14

* feat(cli): add About section

This commit adds an about section to the app that explains it in more
detail. There is also a brief paragraph at the top of the table that
describes how to use/work with it.

Closes #14