docs: add Google API key for Google Maps JavaScript library
This commit adds a Google API key for Google Maps JavaScript library so
those that are interested in quickly getting started with this project
can do so and not have to configure one themselves. Not sure this is a
great idea. Probably will remove the key after having second thoughts.
:)