docs: update README with Install, Quickstart, and About section (#19)
* docs: add Google API key for Google Maps JavaScript library

This commit adds a Google API key for Google Maps JavaScript library so
those that are interested in quickly getting started with this project
can do so and not have to configure one themselves. Not sure this is a
great idea. Probably will remove the key after having second thoughts.
:)

* docs: update README with an About section

Added an About section to the README that explains what the app is, how
it works, and why I built it.

* docs: fix typo in the About section of README

* docs: update README with Install and Quickstart section

Added an Install and  Quickstart section to the README that explains how
to get going with the app right out of the box without having to sign up
for the 3rd party services.

* docs: clarify the Quickstart details